Allow registration with empty interest
In order to temporarily disable event listening for a specific handle,
it may be convenient to reregister it with an empty interest.

This avoids to temporarily deregister the handle before registering it
again (which, by the way, is not supported on Windows[1]).

Therefore, remove the check that prohibited registration with empty
interest.

Note that ERROR and HUP may still be triggered in that case.

[1]: https://github.com/carllerche/mio/issues/633